Added temporary Game Options
This is a popup that is supposed to show before starting a game. It provides the initial values for a game so that we know what to display. Utilizes the Game Schema I committed earlier, which is helpful in case we want to setup a game that is already running.